+++ assets/js/base/components/cart-checkout/totals/fees/index.js
@@ -9,7 +9,7 @@ import PropTypes from 'prop-types';
 /**
  * Internal dependencies
  */
-import TotalsItem from '../totals-item';
+import TotalsItem from '../item';
 
 const TotalsFees = ( { currency, values } ) => {
 	const { needsShipping } = useShippingDataContext();
