+++ assets/js/base/components/cart-checkout/totals/footer-item/index.js
@@ -13,7 +13,7 @@ import PropTypes from 'prop-types';
 /**
  * Internal dependencies
  */
-import TotalsItem from '../totals-item';
+import TotalsItem from '../item';
 import './style.scss';
 
 const SHOW_TAXES = TAXES_ENABLED && DISPLAY_CART_PRICES_INCLUDING_TAX;
