+++ assets/js/base/components/cart-checkout/totals/totals-taxes-item/index.js
@@ -11,7 +11,7 @@ import {
 /**
  * Internal dependencies
  */
-import TotalsItem from '../totals-item';
+import TotalsItem from '../item';
 import './style.scss';
 
 const TotalsTaxesItem = ( { currency, values } ) => {
